Refactor ExamplesCommand and ExamplesInteractiveServer
