+++ core/src/main/kotlin/io/specmatic/core/examples/module/ExampleGenerationStatus.kt
@@ -1,0 +1,3 @@
+package io.specmatic.core.examples.module
+
+enum class ExampleGenerationStatus { CREATED, EXISTED }
