+++ core/src/main/kotlin/io/specmatic/core/examples/server/ExamplesInteractiveServer.kt
@@ -11,40 +11,28 @@ import io.ktor.server.plugins.cors.routing.*
 import io.ktor.server.request.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
-import io.specmatic.conversions.ExampleFromFile
-import io.specmatic.core.*
-import io.specmatic.core.discriminator.DiscriminatorExampleInjector
-import io.specmatic.core.discriminator.DiscriminatorMetadata
-import io.specmatic.core.examples.server.ExamplesView.Companion.isScenarioMultiGen
+import io.specmatic.core.DEFAULT_TIMEOUT_IN_MILLISECONDS
+import io.specmatic.core.Feature
+import io.specmatic.core.Result
+import io.specmatic.core.SPECMATIC_STUB_DICTIONARY
+import io.specmatic.core.examples.module.ExampleFixModule
+import io.specmatic.core.examples.module.ExampleGenerationModule
+import io.specmatic.core.examples.module.ExampleValidationModule
 import io.specmatic.core.examples.server.ExamplesView.Companion.toTableRows
-import io.specmatic.core.examples.server.SchemaExample.Companion.toSchemaExampleFileName
 import io.specmatic.core.examples.server.SchemaExamplesView.Companion.schemaExamplesToTableRows
 import io.specmatic.core.filters.ScenarioMetadataFilter
 import io.specmatic.core.filters.ScenarioMetadataFilter.Companion.filterUsing
-import io.specmatic.core.log.consoleDebug
-import io.specmatic.core.log.consoleLog
 import io.specmatic.core.log.logger
+import io.specmatic.core.parseContractFileToFeature
 import io.specmatic.core.pattern.ContractException
-import io.specmatic.core.pattern.HasFailure
 import io.specmatic.core.route.modules.HealthCheckModule.Companion.configureHealthCheckModule
 import io.specmatic.core.utilities.exceptionCauseMessage
-import io.specmatic.core.utilities.uniqueNameForApiOperation
-import io.specmatic.core.value.JSONArrayValue
-import io.specmatic.core.value.JSONObjectValue
-import io.specmatic.core.value.NullValue
-import io.specmatic.core.value.ScalarValue
-import io.specmatic.core.value.StringValue
-import io.specmatic.core.value.Value
-import io.specmatic.mock.MOCK_HTTP_REQUEST
-import io.specmatic.mock.MOCK_HTTP_RESPONSE
-import io.specmatic.mock.ScenarioStub
 import io.specmatic.test.ContractTest
 import io.specmatic.test.TestInteractionsLog
 import io.specmatic.test.TestInteractionsLog.combineLog
 import java.io.Closeable
 import java.io.File
 import java.io.FileNotFoundException
-import java.util.concurrent.atomic.AtomicInteger
 import kotlin.system.exitProcess
 
 class ExamplesInteractiveServer(
@@ -59,6 +47,11 @@ class ExamplesInteractiveServer(
     private val allowOnlyMandatoryKeysInJSONObject: Boolean
 ) : Closeable {
     private var contractFileFromRequest: File? = null
+
+    private val exampleModule = ExampleModule()
+    private val exampleGenerationModule = ExampleGenerationModule(exampleModule)
+    private val exampleValidationModule = ExampleValidationModule()
+    private val exampleFixModule = ExampleFixModule(exampleValidationModule)
 
     init {
         if(externalDictionaryFile != null) System.setProperty(SPECMATIC_STUB_DICTIONARY, externalDictionaryFile.path)
@@ -133,9 +126,9 @@ class ExamplesInteractiveServer(
                     try {
                         val request = call.receive<GenerateExampleRequest>()
                         val generatedExamples = if (request.isSchemaBased) {
-                            generateForSchemaBased(contractFile, request.path, request.method)
+                            exampleGenerationModule.generateForSchemaBased(contractFile, request.path, request.method)
                         } else {
-                            generate(
+                            exampleGenerationModule.generate(
                                 contractFile,
                                 request.method,
                                 request.path,
@@ -157,7 +150,7 @@ class ExamplesInteractiveServer(
                     try {
                         val contractFile = getContractFile()
                         val validationResultResponse = try {
-                            val result = validateExample(contractFile, File(request.exampleFile))
+                            val result = exampleValidationModule.validateExample(contractFile, File(request.exampleFile))
                             if (result is Result.Failure) {
                                 ValidateExampleResponse(
                                     absPath = request.exampleFile, errorMessage = result.reportString(),
@@ -188,7 +181,7 @@ class ExamplesInteractiveServer(
                     try {
                         val contractFile = getContractFile()
                         val fixExamplesResponse = try {
-                            fixExample(contractFile, request)
+                            exampleFixModule.fixExample(contractFile, request)
                         } catch (e: FileNotFoundException) {
                             FixExampleResponse(request.exampleFile, e.message ?: "File not found")
                         } catch (e: ContractException) {
@@ -286,9 +279,9 @@ class ExamplesInteractiveServer(
             filter
         ).filter(parseContractFileToFeature(contractFile))
 
-        val examplesDir = getExamplesDirPath(contractFile)
+        val examplesDir = exampleModule.getExamplesDirPath(contractFile)
         val endpoints = ExamplesView.getEndpoints(feature, examplesDir)
-        val schemaExamplesPairs = examplesDir.getSchemaExamplesWithValidation(feature)
+        val schemaExamplesPairs = exampleModule.getSchemaExamplesWithValidation(feature, examplesDir)
 
         val exampleTableRows = endpoints.toTableRows()
         val schemaTableRows = schemaExamplesToTableRows(feature, schemaExamplesPairs)
@@ -324,589 +317,44 @@ class ExamplesInteractiveServer(
         }
     }
 
-    class ScenarioFilter(filterName: String = "", filterNotName: String = "", filterClauses: String = "") {
-        private val filter = filterClauses
+    private fun testExample(test: ContractTest, testBaseUrl: String): Pair<Result, String> {
+        val testResult = test.runTest(testBaseUrl, timeoutInMilliseconds = DEFAULT_TIMEOUT_IN_MILLISECONDS)
+        val testLog = TestInteractionsLog.testHttpLogMessages.lastOrNull { it.scenario == testResult.first.scenario }
 
-        private val filterNameTokens = if(filterName.isNotBlank()) {
-            filterName.trim().split(",").map { it.trim() }
-        } else emptyList()
-
-        private val filterNotNameTokens = if(filterNotName.isNotBlank()) {
-            filterNotName.trim().split(",").map { it.trim() }
-        } else emptyList()
-
-        fun filter(feature: Feature): Feature {
-            val scenariosFilteredByOlderSyntax = feature.scenarios.filter { scenario ->
-                if(filterNameTokens.isNotEmpty()) {
-                    filterNameTokens.any { name -> scenario.testDescription().contains(name) }
-                } else true
-            }.filter { scenario ->
-                if(filterNotNameTokens.isNotEmpty()) {
-                    filterNotNameTokens.none { name -> scenario.testDescription().contains(name) }
-                } else true
-            }
-
-            val scenarioFilter = ScenarioMetadataFilter.from(filter)
-
-            val filteredScenarios = filterUsing(scenariosFilteredByOlderSyntax.asSequence(), scenarioFilter) {
-                it.toScenarioMetadata()
-            }.toList()
-
-
-            return feature.copy(scenarios = filteredScenarios)
-        }
+        return testResult.first to (testLog?.combineLog() ?: "No Test Logs Found")
     }
+}
 
-    companion object {
-        private val exampleFileNamePostFixCounter = AtomicInteger(0)
-        fun resetExampleFileNameCounter() {
-            exampleFileNamePostFixCounter.set(0)
+class ScenarioFilter(filterName: String = "", filterNotName: String = "", filterClauses: String = "") {
+    private val filter = filterClauses
+
+    private val filterNameTokens = if(filterName.isNotBlank()) {
+        filterName.trim().split(",").map { it.trim() }
+    } else emptyList()
+
+    private val filterNotNameTokens = if(filterNotName.isNotBlank()) {
+        filterNotName.trim().split(",").map { it.trim() }
+    } else emptyList()
+
+    fun filter(feature: Feature): Feature {
+        val scenariosFilteredByOlderSyntax = feature.scenarios.filter { scenario ->
+            if(filterNameTokens.isNotEmpty()) {
+                filterNameTokens.any { name -> scenario.testDescription().contains(name) }
+            } else true
+        }.filter { scenario ->
+            if(filterNotNameTokens.isNotEmpty()) {
+                filterNotNameTokens.none { name -> scenario.testDescription().contains(name) }
+            } else true
         }
 
-        enum class ExampleGenerationStatus { CREATED, EXISTED }
-        data class ExamplePathInfo(val path: String, val created: Boolean, val status: ExampleGenerationStatus) {
-            constructor(path: String, created: Boolean): this(path, created, if (created) ExampleGenerationStatus.CREATED else ExampleGenerationStatus.EXISTED)
-            fun relativeTo(file: File): String {
-                return File(path).canonicalFile.relativeTo(file.canonicalFile.parentFile).path
-            }
-        }
+        val scenarioFilter = ScenarioMetadataFilter.from(filter)
 
-        fun testExample(test: ContractTest, testBaseUrl: String): Pair<Result, String> {
-            val testResult = test.runTest(testBaseUrl, timeoutInMilliseconds = DEFAULT_TIMEOUT_IN_MILLISECONDS)
-            val testLog = TestInteractionsLog.testHttpLogMessages.lastOrNull { it.scenario == testResult.first.scenario }
+        val filteredScenarios = filterUsing(scenariosFilteredByOlderSyntax.asSequence(), scenarioFilter) {
+            it.toScenarioMetadata()
+        }.toList()
 
-            return testResult.first to (testLog?.combineLog() ?: "No Test Logs Found")
-        }
 
-        fun generate(contractFile: File, scenarioFilter: ScenarioFilter, extensive: Boolean, allowOnlyMandatoryKeysInJSONObject: Boolean): List<String> {
-            return try {
-                val feature: Feature = parseContractFileToFeature(contractFile).let { feature ->
-                    val filteredScenarios = if (!extensive) {
-                        feature.scenarios.filter {
-                            it.status.toString().startsWith("2")
-                        }
-                    } else {
-                        feature.scenarios
-                    }
-
-                    scenarioFilter.filter(feature.copy(scenarios = filteredScenarios.map {
-                        it.copy(examples = emptyList())
-                    })).copy(stubsFromExamples = emptyMap())
-                }
-                if (feature.scenarios.isEmpty()) {
-                    logger.log("All examples were filtered out by the filter expression")
-                    return emptyList()
-                }
-
-                val examplesDir = getExamplesDirPath(contractFile).also { if (it.exists()) it.mkdirs() }
-                val allExistingExamples = examplesDir.getExamplesFromDir()
-
-                val schemaExamples = schemaExamplesToTableRows(feature, examplesDir.getSchemaExamplesWithValidation(feature)).flatMap {
-                    if (it.example != null) {
-                        listOf(ExamplePathInfo(path = it.example, created = false, status = ExampleGenerationStatus.EXISTED))
-                    } else generateForSchemaBased(contractFile, it.rawPath, it.method)
-                }
-
-                val allExamples = feature.scenarios.flatMap { scenario ->
-                    generateAndLogScenarioExamples(contractFile, feature, scenario, allExistingExamples, allowOnlyMandatoryKeysInJSONObject)
-                }.plus(schemaExamples)
-
-                generationSummary(contractFile, allExamples).map { it.path }
-            } catch (e: StackOverflowError) {
-                logger.log("Got a stack overflow error. You probably have a recursive data structure definition in the contract.")
-                throw e
-            }
-        }
-
-        private fun generateAndLogScenarioExamples(contractFile: File, feature: Feature, scenario: Scenario, allExistingExamples: List<ExampleFromFile>, allowOnlyMandatoryKeysInJSONObject: Boolean): List<ExamplePathInfo> {
-            return try {
-                val trimmedDescription = scenario.testDescription().trim()
-                val existingExamples = getExistingExampleFiles(feature, scenario, allExistingExamples).map { it.first }
-                val scenarioIsMultiGen = isScenarioMultiGen(scenario, scenario.resolver)
-
-                val generatedExamples = if (scenarioIsMultiGen || existingExamples.isEmpty()) {
-                    generateExampleFiles(contractFile, feature, scenario, allowOnlyMandatoryKeysInJSONObject, existingExamples)
-                } else emptyList()
-
-                val allExamples = existingExamples.map {
-                    ExamplePathInfo(it.file.canonicalPath, created = false)
-                }.plus(generatedExamples)
-
-                allExamples.also { logExamples(contractFile, trimmedDescription, it) }
-            } catch (e: Throwable) {
-                logger.log(e, "Exception generating example for ${scenario.testDescription()}")
-                emptyList()
-            }
-        }
-
-        private fun logExamples(contractFile: File, description: String, examples: List<ExamplePathInfo>) {
-            examples.forEach { example ->
-                val loggablePath = example.relativeTo(contractFile)
-                if (!example.created) {
-                    consoleLog("Example already existed for $description: $loggablePath")
-                } else {
-                    consoleLog("Created example for $description: $loggablePath")
-                }
-            }
-        }
-
-        private fun generationSummary(contractFile: File, exampleFiles: List<ExamplePathInfo>): List<ExamplePathInfo> {
-            val resultCounts = exampleFiles.groupBy { it.status }.mapValues { it.value.size }
-            val createdFileCount = resultCounts[ExampleGenerationStatus.CREATED] ?: 0
-            val existingFileCount = resultCounts[ExampleGenerationStatus.EXISTED] ?: 0
-
-            logger.log(System.lineSeparator() + "NOTE: All examples may be found in ${getExamplesDirPath(contractFile).canonicalFile}" + System.lineSeparator())
-            logger.log("=============== Example Generation Summary ===============")
-            logger.log("$createdFileCount example(s) created, $existingFileCount examples already existed")
-            logger.log("==========================================================")
-
-            return exampleFiles
-        }
-
-        fun generate(
-            contractFile: File,
-            method: String,
-            path: String,
-            responseStatusCode: Int,
-            contentType: String? = null,
-            bulkMode: Boolean = false,
-            allowOnlyMandatoryKeysInJSONObject: Boolean
-        ): List<ExamplePathInfo> {
-            val feature = parseContractFileToFeature(contractFile)
-            val scenario = feature.scenarioAssociatedTo(method, path, responseStatusCode, contentType)
-            if(scenario == null) return emptyList()
-
-            val examplesDir = getExamplesDirPath(contractFile)
-            val examples = examplesDir.getExamplesFromDir()
-
-            return generateAndLogScenarioExamples(
-                contractFile, feature, scenario,
-                allExistingExamples = if(bulkMode) examples else emptyList(),
-                allowOnlyMandatoryKeysInJSONObject
-            )
-        }
-
-        fun fixExample(contractFile: File, request: FixExampleRequest): FixExampleResponse {
-            try {
-                val feature = parseContractFileToFeature(contractFile)
-                fixExample(feature, request.exampleFile)
-                return FixExampleResponse(exampleFile = request.exampleFile)
-            } catch(e: Throwable) {
-                return FixExampleResponse(exampleFile = request.exampleFile, errorMessage = exceptionCauseMessage(e))
-            }
-        }
-
-        fun fixExample(feature: Feature, exampleFile: File): FixExampleResult {
-            val exampleReturnValue = ExampleFromFile.fromFile(exampleFile)
-
-            if(exampleReturnValue is HasFailure<ExampleFromFile>) {
-                if(validateSchemaExample(feature, exampleFile) is Result.Success) {
-                    return FixExampleResult(status = FixExampleStatus.SKIPPED, exampleFileName = exampleFile.name)
-                }
-                fixSchemaExampleAndWriteTo(exampleFile, feature)
-                return FixExampleResult(status = FixExampleStatus.SUCCEDED, exampleFileName = exampleFile.name)
-            }
-
-            val example = exampleReturnValue.value
-            val matchingHttpPathPattern = feature.matchingHttpPathPatternFor(
-                example.requestPath.orEmpty()
-            ) ?: throw Exception("No scenario found for request path in '${exampleFile.name}'.")
-
-            val scenario = feature.scenarioAssociatedTo(
-                method = example.requestMethod.orEmpty(),
-                path = matchingHttpPathPattern.path,
-                responseStatusCode = example.responseStatus ?: 0,
-                contentType = example.requestContentType
-            ) ?: throw Exception("No scenario found for example '${exampleFile.name}'.")
-
-            if(validateExample(feature, exampleFile) is Result.Success) {
-                return FixExampleResult(status = FixExampleStatus.SKIPPED, exampleFileName = exampleFile.name)
-            }
-
-            fixExampleAndWriteTo(exampleFile, scenario, feature)
-            return FixExampleResult(status = FixExampleStatus.SUCCEDED, exampleFileName = exampleFile.name)
-        }
-
-        private fun fixExampleAndWriteTo(exampleFile: File, scenario: Scenario, feature: Feature) {
-            val example = ScenarioStub.readFromFile(exampleFile)
-            val (fixedRequest, fixedResponse) = scenario.fixRequestResponse(
-                httpRequest = example.request,
-                httpResponse = example.response,
-                flagsBased = feature.flagsBased
-            )
-            val fixedExampleJson = example.copy(
-                request = fixedRequest,
-                response = fixedResponse
-            ).toJSON().toStringLiteral()
-
-            exampleFile.writeText(fixedExampleJson)
-        }
-
-        private fun fixSchemaExampleAndWriteTo(exampleFile: File, feature: Feature) {
-            val schemaExample = SchemaExample.fromFile(exampleFile).value
-            val fixedExample = feature.fixSchemaFlagBased(
-                schemaExample.discriminatorBasedOn,
-                schemaExample.schemaBasedOn,
-                schemaExample.value
-            )
-            schemaExample.file.writeText(fixedExample.toStringLiteral())
-        }
-
-        fun generateForSchemaBased(contractFile: File, path: String, method: String): List<ExamplePathInfo> {
-            val examplesDir = getExamplesDirPath(contractFile)
-            if(examplesDir.exists().not()) examplesDir.mkdirs()
-
-            val feature = parseContractFileToFeature(contractFile)
-            val (discriminatorPatternName, patternName) = when {
-                method.isEmpty() -> null to path
-                else -> path to method
-            }
-
-            val value = feature.generateSchemaFlagBased(discriminatorPatternName, patternName)
-            val schemaFileName = toSchemaExampleFileName(discriminatorPatternName, patternName)
-
-            val exampleFile = examplesDir.getSchemaExamples().firstOrNull {
-                it.file.nameWithoutExtension == schemaFileName
-            }?.file ?: examplesDir.resolve(schemaFileName)
-
-            println("Writing to file: ${exampleFile.relativeTo(contractFile.canonicalFile.parentFile).path}")
-            exampleFile.writeText(value.toStringLiteral())
-            return listOf(ExamplePathInfo(path = exampleFile.absolutePath, created = true))
-        }
-
-        private fun generateExampleFiles(
-            contractFile: File,
-            feature: Feature,
-            scenario: Scenario,
-            allowOnlyMandatoryKeysInJSONObject: Boolean,
-            existingExamples: List<ExampleFromFile>
-        ): List<ExamplePathInfo> {
-            val examplesDir = getExamplesDirPath(contractFile)
-            if(!examplesDir.exists()) examplesDir.mkdirs()
-
-            val discriminatorBasedRequestResponses = feature
-                .generateDiscriminatorBasedRequestResponseList(
-                    scenario,
-                    allowOnlyMandatoryKeysInJSONObject = allowOnlyMandatoryKeysInJSONObject
-                ).map { it.copy(response = it.response.cleanup()) }
-
-            val requestDiscriminator = discriminatorBasedRequestResponses.first().requestDiscriminator
-            val responseDiscriminator = discriminatorBasedRequestResponses.first().responseDiscriminator
-
-            val existingDiscriminators = existingExamples.map {
-                it.requestBody?.getDiscriminatorValue(requestDiscriminator).orEmpty() to it.responseBody?.getDiscriminatorValue(responseDiscriminator).orEmpty()
-            }.toSet()
-
-            return discriminatorBasedRequestResponses.filterNot { it.matches(existingDiscriminators) }.map { (request, response, requestDiscriminator, responseDiscriminator) ->
-                val requestWithoutAttrSelection = request.removeAttrSelection(scenario.attributeSelectionPattern)
-
-                val scenarioStub = ScenarioStub(requestWithoutAttrSelection, response)
-                val jsonWithDiscriminator = DiscriminatorExampleInjector(
-                    stubJSON = scenarioStub.toJSON(),
-                    requestDiscriminator = requestDiscriminator,
-                    responseDiscriminator = responseDiscriminator
-                ).getExampleWithDiscriminator()
-
-                val uniqueNameForApiOperation = getExampleFileNameBasedOn(
-                    requestDiscriminator,
-                    responseDiscriminator,
-                    scenarioStub
-                )
-
-                val file = examplesDir.resolve("${uniqueNameForApiOperation}_${exampleFileNamePostFixCounter.incrementAndGet()}.json")
-                println("Writing to file: ${file.relativeTo(contractFile.canonicalFile.parentFile).path}")
-                file.writeText(jsonWithDiscriminator.toStringLiteral())
-                ExamplePathInfo(file.absolutePath, true)
-            }
-        }
-
-        private fun HttpRequest.removeAttrSelection(attributeSelectionPattern: AttributeSelectionPatternDetails): HttpRequest {
-            return this.copy(
-                queryParams = this.queryParams.remove(attributeSelectionPattern.getQueryParamKey())
-            )
-        }
-
-        private fun Value.getDiscriminatorValue(discriminator: DiscriminatorMetadata): String? {
-            return when (this) {
-                is JSONObjectValue -> {
-                    val targetValue = this.getEventValue() ?: this
-                    targetValue.findFirstChildByPath(discriminator.discriminatorProperty)?.toStringLiteral()
-                }
-                is JSONArrayValue -> this.list.first().getDiscriminatorValue(discriminator)
-                else -> null
-            }
-        }
-
-        private fun JSONObjectValue.getEventValue(): JSONObjectValue? {
-            return (this.findFirstChildByPath("event") as? JSONObjectValue)?.let { eventValue ->
-                eventValue.findFirstChildByPath(eventValue.jsonObject.keys.first()) as? JSONObjectValue
-            }
-        }
-
-        private fun DiscriminatorBasedRequestResponse.matches(discriminatorValues: Set<Pair<String, String>>): Boolean {
-            return discriminatorValues.contains(requestDiscriminator.discriminatorValue to responseDiscriminator.discriminatorValue)
-        }
-
-        fun validateInlineExamples(
-            feature: Feature,
-            examples: Map<String, List<ScenarioStub>> = emptyMap(),
-            scenarioFilter: ScenarioFilter = ScenarioFilter()
-        ): Map<String, Result> {
-            val updatedFeature = scenarioFilter.filter(feature)
-
-            val results = examples.mapValues { (name, exampleList) ->
-                logger.debug("Validating $name")
-
-                exampleList.mapNotNull { example ->
-                    val results = validateExample(updatedFeature, example)
-                    if (!results.hasResults()) return@mapNotNull null else results.toResultIfAny()
-                }.let {
-                    Result.fromResults(it)
-                }
-            }
-
-            return results
-        }
-
-        fun validateExamples(
-            feature: Feature,
-            examples: List<File> = emptyList(),
-            scenarioFilter: ScenarioFilter = ScenarioFilter()
-        ): Map<String, Result> {
-            val updatedFeature = scenarioFilter.filter(feature)
-
-            val results = examples.associate { exampleFile ->
-                logger.debug("Validating ${exampleFile.name}")
-                exampleFile.canonicalPath to validateExample(updatedFeature, exampleFile)
-            }
-
-            return results
-        }
-
-        fun validateExample(contractFile: File, exampleFile: File): Result {
-            val feature = parseContractFileToFeature(contractFile)
-            return validateExample(feature, exampleFile)
-        }
-
-        fun validateExample(feature: Feature, scenarioStub: ScenarioStub): Results {
-            return feature.matchResultFlagBased(scenarioStub, InteractiveExamplesMismatchMessages)
-        }
-
-        private fun validateExample(feature: Feature, example: ExampleFromFile): Result {
-            return feature.matchResultFlagBased(example.request, example.response, InteractiveExamplesMismatchMessages).toResultIfAnyWithCauses()
-        }
-
-        private fun validateExample(feature: Feature, schemaExample: SchemaExample): Result {
-            if (schemaExample.value is NullValue) {
-                return Result.Success()
-            }
-
-            return feature.matchResultSchemaFlagBased(
-                discriminatorPatternName = schemaExample.discriminatorBasedOn,
-                patternName = schemaExample.schemaBasedOn,
-                value = schemaExample.value,
-                mismatchMessages = InteractiveExamplesMismatchMessages,
-                breadCrumbIfDiscriminatorMismatch = schemaExample.file.name
-            )
-        }
-
-        private fun validateExample(feature: Feature, exampleFile: File): Result {
-            return ExampleFromFile.fromFile(exampleFile).realise(
-                hasValue = { example, _ -> validateExample(feature, example) },
-                orFailure = { validateSchemaExample(feature, exampleFile) },
-                orException = { it.toHasFailure().failure }
-            )
-        }
-
-        private fun validateSchemaExample(feature: Feature, exampleFile: File): Result {
-            return SchemaExample.fromFile(exampleFile).realise(
-                hasValue = { example, _ -> validateExample(feature, example) },
-                orException = { it.toHasFailure().failure },
-                orFailure = { it.failure }
-            )
-        }
-
-        private fun HttpResponse.cleanup(): HttpResponse {
-            return this.copy(headers = this.headers.minus(SPECMATIC_RESULT_HEADER))
-        }
-
-        fun getExistingExampleFiles(feature: Feature, scenario: Scenario, examples: List<ExampleFromFile>): List<Pair<ExampleFromFile, Result>> {
-            return examples.mapNotNull { example ->
-                when (val matchResult = scenario.matches(example.request, example.response, InteractiveExamplesMismatchMessages, feature.flagsBased)) {
-                    is Result.Success -> example to matchResult
-                    is Result.Failure -> {
-                        val isFailureRelatedToScenario = matchResult.getFailureBreadCrumbs("").none { breadCrumb ->
-                            breadCrumb.contains(PATH_BREAD_CRUMB)
-                                    || breadCrumb.contains(METHOD_BREAD_CRUMB)
-                                    || breadCrumb.contains("REQUEST.HEADERS.Content-Type")
-                                    || breadCrumb.contains("STATUS")
-                        }
-                        if (isFailureRelatedToScenario) { example to matchResult } else null
-                    }
-                }
-            }
-        }
-
-        fun getExamplesDirPath(contractFile: File): File {
-            return contractFile.canonicalFile
-                .parentFile
-                .resolve("""${contractFile.nameWithoutExtension}$EXAMPLES_DIR_SUFFIX""")
-        }
-
-        fun File.getExamplesFromDir(): List<ExampleFromFile> {
-            return this.listFiles().orEmpty().filter { it.extension == "json" }.mapNotNull {
-                ExampleFromFile.fromFile(it).realise(
-                    hasValue = { example, _ -> example },
-                    orException = { err -> consoleDebug(exceptionCauseMessage(err.t)); null },
-                    orFailure = { null }
-                )
-            }
-        }
-
-        fun File.getSchemaExamplesWithValidation(feature: Feature): List<Pair<SchemaExample, Result?>> {
-            return getSchemaExamples().map {
-                it to if(it.value !is NullValue) {
-                    feature.matchResultSchemaFlagBased(
-                        discriminatorPatternName = it.discriminatorBasedOn,
-                        patternName = it.schemaBasedOn,
-                        value = it.value,
-                        mismatchMessages = InteractiveExamplesMismatchMessages,
-                        breadCrumbIfDiscriminatorMismatch = it.file.name
-                    )
-                } else null
-            }
-        }
-
-        private fun File.getSchemaExamples(): List<SchemaExample> {
-            return this.listFiles().orEmpty().filter { it.extension == "json" }.mapNotNull {
-                SchemaExample.fromFile(it).realise(
-                    hasValue = { example, _ -> example },
-                    orException = { err -> consoleDebug(exceptionCauseMessage(err.t)); null },
-                    orFailure = { null }
-                )
-            }
-        }
-
-        private fun getExampleFileNameBasedOn(
-            requestDiscriminator: DiscriminatorMetadata,
-            responseDiscriminator: DiscriminatorMetadata,
-            scenarioStub: ScenarioStub
-        ): String {
-            val discriminatorValue = requestDiscriminator.discriminatorValue.ifBlank {
-                responseDiscriminator.discriminatorValue
-            }
-            val discriminatorName = if (discriminatorValue.isNotEmpty()) "${discriminatorValue}_" else ""
-            return discriminatorName + uniqueNameForApiOperation(
-                scenarioStub.request,
-                "",
-                scenarioStub.response.status
-            )
-        }
-
-        private fun Value.isScalarOrEmpty(): Boolean {
-            return this is ScalarValue || this is NoBodyValue
-        }
-
-        private fun Value.insertFieldsInValue(fieldsToBeInserted: Map<String, Value>): Value {
-            return when (this) {
-                is JSONObjectValue -> JSONObjectValue(fieldsToBeInserted.plus(this.jsonObject))
-                is JSONArrayValue -> JSONArrayValue(this.list.map {value ->  value.insertFieldsInValue(fieldsToBeInserted) })
-                else -> this
-            }
-        }
-
-        private fun Map<String, String>.toValueMap(): Map<String, Value> {
-            return this.mapValues { StringValue(it.value) }
-        }
-
-        fun transformExistingExamples(contractFile: File, overlayFile: File?, examplesDir: File) {
-            val feature = parseContractFileToFeature(contractPath = contractFile.absolutePath, overlayContent = overlayFile?.readText().orEmpty())
-            val examples = examplesDir.getExamplesFromDir()
-
-            examples.forEach { example ->
-                consoleDebug("\nTransforming ${example.file.nameWithoutExtension}")
-
-                if (example.request.body.isScalarOrEmpty() && example.response.body.isScalarOrEmpty()) {
-                    consoleDebug("Skipping ${example.file.name}, both request and response bodies are scalars")
-                    return@forEach
-                }
-
-                val scenario = feature.matchResultFlagBased(example.request, example.response, InteractiveExamplesMismatchMessages)
-                    .toResultIfAny().takeIf { it.isSuccess() }?.scenario as? Scenario
-
-                if (scenario == null) {
-                    consoleDebug("Skipping ${example.file.name}, no matching scenario found")
-                    return@forEach
-                }
-
-                val flagBasedResolver = feature.flagsBased.update(scenario.resolver)
-                val requestWithoutOptionality = scenario.httpRequestPattern.withoutOptionality(example.request, flagBasedResolver)
-                val responseWithoutOptionality = scenario.httpResponsePattern.withoutOptionality(example.response, flagBasedResolver)
-
-                val updatedExample = example.replaceWithDescriptions(requestWithoutOptionality, responseWithoutOptionality)
-                consoleDebug("Writing transformed example to ${example.file.canonicalFile.relativeTo(contractFile).path}")
-                example.file.writeText(updatedExample.toStringLiteral())
-                consoleDebug("Successfully written transformed example")
-            }
-        }
-
-        private fun ExampleFromFile.replaceWithDescriptions(request: HttpRequest, response: HttpResponse): JSONObjectValue {
-            return this.json.jsonObject.mapValues { (key, value) ->
-                when (key) {
-                    MOCK_HTTP_REQUEST -> request.toJSON().insertFieldsInValue(value.getDescriptionMap())
-                    MOCK_HTTP_RESPONSE -> response.toJSON().insertFieldsInValue(value.getDescriptionMap())
-                    else -> value
-                }
-            }.let { JSONObjectValue(it.toMap()) }
-        }
-
-        private fun Value.getDescriptionMap(): Map<String, Value> {
-            return (this as? JSONObjectValue)
-                ?.findFirstChildByPath("description")
-                ?.let { mapOf("description" to it.toStringLiteral()).toValueMap() }
-                ?: emptyMap()
-        }
-
-        fun externaliseInlineExamples(contractFile: File): File {
-            val feature = parseContractFileToFeature(contractFile)
-            val inlineStubs: List<ScenarioStub> = feature.stubsFromExamples.flatMap {
-                it.value.map { (request, response) -> ScenarioStub(request, response) }
-            }
-            try {
-                inlineStubs.forEach { writeToExampleFile(it, contractFile) }
-            } catch(e: Exception) {
-                consoleLog(e)
-            }
-            return getExamplesDirPath(contractFile)
-        }
-
-        private fun writeToExampleFile(scenarioStub: ScenarioStub, contractFile: File): File {
-            val examplesDir = getExamplesDirPath(contractFile)
-            if (examplesDir.exists().not()) examplesDir.mkdirs()
-            val stubJSON = scenarioStub.toJSON()
-            val uniqueNameForApiOperation =
-                uniqueNameForApiOperation(scenarioStub.request, "", scenarioStub.response.status)
-
-            val file =
-                examplesDir.resolve("${uniqueNameForApiOperation}_${exampleFileNamePostFixCounter.incrementAndGet()}.json")
-            println("Writing to file: ${file.relativeTo(contractFile.canonicalFile.parentFile).path}")
-            file.writeText(stubJSON.toStringLiteral())
-            return file
-        }
-        private fun validatePartialFailures(contractFile: File, exampleFile: File): Any {
-            val feature = parseContractFileToFeature(contractFile)
-            val exampleFromFile = ExampleFromFile.fromFile(exampleFile).realise(
-                hasValue = { example, _ -> validatePartialFailures(feature, example) },
-                orFailure = { it.failure },
-                orException = { it.toHasFailure().failure }
-            )
-            return exampleFromFile
-        }
-
-        private fun validatePartialFailures(feature: Feature, exampleFromFile: ExampleFromFile): List<Result> {
-            return feature.matchResultFlagBased(exampleFromFile.request, exampleFromFile.response, InteractiveExamplesMismatchMessages).toResultPartialFailures()
-        }
-
+        return feature.copy(scenarios = filteredScenarios)
     }
 }
