+++ core/src/main/kotlin/io/specmatic/core/examples/server/GenerateExampleResponse.kt
@@ -1,10 +1,12 @@
 package io.specmatic.core.examples.server
 
+import io.specmatic.core.examples.module.ExamplePathInfo
+
 data class GenerateExampleResponse (
     val examples: List<GenerateExample>
 ) {
     companion object {
-        fun from(infos: List<ExamplesInteractiveServer.Companion.ExamplePathInfo>): GenerateExampleResponse {
+        fun from(infos: List<ExamplePathInfo>): GenerateExampleResponse {
             return GenerateExampleResponse(infos.map { GenerateExample(it.path, it.created) })
         }
     }
