+++ core/src/test/kotlin/io/specmatic/core/examples/module/ExampleTransformationModuleTest.kt
@@ -1,0 +1,3 @@
+package io.specmatic.core.examples.module
+
+class ExampleTransformationModuleTest
